#CLIB-57 - added missing method implementation - cleanup - coding guidelines fixes
git-svn-id: http://escidev5.fiz-karlsruhe.de:8888/prototype/branches/eSciDocCoreClient/1.2/eSciDocCoreClient-rest/eSciDocCoreClient/@39000 5c398c58-3d0a-0410-a38c-d0825f6816e0

+++ src/main/java/de/escidoc/core/client/OrganizationalUnitHandlerClient.java
@@ -44,6 +44,8 @@ import de.escidoc.core.client.interfaces.OrganizationalUnitHandlerClientInterfac
 import de.escidoc.core.client.rest.RestOrganizationalUnitHandlerClient;
 import de.escidoc.core.client.soap.SoapOrganizationalUnitHandlerClient;
 import de.escidoc.core.common.jibx.Factory;
+import de.escidoc.core.common.jibx.Marshaller;
+import de.escidoc.core.common.jibx.MarshallerFactory;
 import de.escidoc.core.resources.common.Result;
 import de.escidoc.core.resources.common.TaskParam;
 import de.escidoc.core.resources.common.properties.Properties;
@@ -75,7 +77,7 @@ public class OrganizationalUnitHandlerClient
     public OrganizationalUnitHandlerClient() {
         super();
     }
-    
+
     /**
      * 
      * @param serviceAddress
@@ -83,7 +85,7 @@ public class OrganizationalUnitHandlerClient
     public OrganizationalUnitHandlerClient(final String serviceAddress) {
         super(serviceAddress);
     }
-    
+
     /**
      * See Interface for functional description.
      * 
@@ -92,24 +94,22 @@ public class OrganizationalUnitHandlerClient
      * @throws EscidocClientException
      * @see de.escidoc.core.client.interfaces.OrganizationalUnitHandlerClientInterface#create(de.escidoc.core.resources.interfaces.organizationalUnit.OrganizationalUnitInterface)
      */
+    @Override
     public OrganizationalUnit create(final OrganizationalUnit organizationalUnit)
         throws EscidocException, InternalClientException, TransportException {
 
-        String orgUnitString =
-            Factory
-                .getMarshallerFactory(getTransport())
-                .getOrganizationalUnitMarshaller()
-                .marshalDocument(organizationalUnit);
-        String xml = null;
+        if (organizationalUnit == null)
+            throw new IllegalArgumentException(
+                "organizationalUnit must not be null.");
+
+        String xml = getMarshaller().marshalDocument(organizationalUnit);
         if (getTransport() == TransportProtocol.SOAP) {
-            xml = getSoapHandlerClient().create(orgUnitString);
+            xml = getSoapHandlerClient().create(xml);
         }
         else {
-            xml = getRestHandlerClient().create(orgUnitString);
+            xml = getRestHandlerClient().create(xml);
         }
-        return Factory
-            .getMarshallerFactory(getTransport())
-            .getOrganizationalUnitMarshaller().unmarshalDocument(xml);
+        return getMarshaller().unmarshalDocument(xml);
     }
 
     /**
@@ -120,8 +120,12 @@ public class OrganizationalUnitHandlerClient
      * @throws EscidocClientException
      * @see de.escidoc.core.client.interfaces.OrganizationalUnitHandlerClientInterface#retrieve(java.lang.String)
      */
+    @Override
     public OrganizationalUnit retrieve(final String id)
         throws EscidocException, InternalClientException, TransportException {
+
+        if (id == null)
+            throw new IllegalArgumentException("id must not be null.");
 
         String xml = null;
         if (getTransport() == TransportProtocol.SOAP) {
@@ -130,9 +134,7 @@ public class OrganizationalUnitHandlerClient
         else {
             xml = getRestHandlerClient().retrieve(id);
         }
-        return Factory
-            .getMarshallerFactory(getTransport())
-            .getOrganizationalUnitMarshaller().unmarshalDocument(xml);
+        return getMarshaller().unmarshalDocument(xml);
     }
 
     /**
@@ -145,28 +147,65 @@ public class OrganizationalUnitHandlerClient
      * @throws TransportException
      * @see de.escidoc.core.client.interfaces.CrudHandlerInterface#update(java.lang.Object)
      */
+    @Override
     public OrganizationalUnit update(final OrganizationalUnit organizationalUnit)
         throws EscidocException, InternalClientException, TransportException {
 
-        String xml = null;
-        String orgUnitString =
-            Factory
-                .getMarshallerFactory(getTransport())
-                .getOrganizationalUnitMarshaller()
-                .marshalDocument(organizationalUnit);
+        if (organizationalUnit == null)
+            throw new IllegalArgumentException(
+                "organizationalUnit must not be null.");
+
+        String xml = getMarshaller().marshalDocument(organizationalUnit);
         if (getTransport() == TransportProtocol.SOAP) {
             xml =
                 getSoapHandlerClient().update(organizationalUnit.getObjid(),
-                    orgUnitString);
+                    xml);
         }
         else {
             xml =
                 getRestHandlerClient().update(organizationalUnit.getObjid(),
-                    orgUnitString);
+                    xml);
         }
-        return Factory
-            .getMarshallerFactory(getTransport())
-            .getOrganizationalUnitMarshaller().unmarshalDocument(xml);
+        return getMarshaller().unmarshalDocument(xml);
+    }
+
+    /**
+     * Updates the parents of an Organizational Unit.<br/>
+     * <br/>
+     * Preconditions:
+     * <ul>
+     * <li>The Organizational Unit must exist.</li>
+     * <li>The public-status is "opened".</li>
+     * </ul>
+     * 
+     * @param id
+     *            The identifier of the Organizational Unit.
+     * @param parents
+     *            The Parents object of the corresponding Organizational Unit.
+     * @return The updated Parents object of the Organizational Unit.
+     * @throws EscidocException
+     * @throws InternalClientException
+     * @throws TransportException
+     */
+    public Parents updateParents(final String id, final Parents parents)
+        throws EscidocException, InternalClientException, TransportException {
+        if (id == null)
+            throw new IllegalArgumentException("id must not be null.");
+        if (parents == null)
+            throw new IllegalArgumentException("parents must not be null.");
+
+        Marshaller<Parents> m =
+            Factory.getMarshallerFactory(getTransport()).getMarshaller(
+                MarshallerFactory.CLASS_PARENTS);
+        String xml = m.marshalDocument(parents);
+
+        if (getTransport() == TransportProtocol.SOAP) {
+            xml = getSoapHandlerClient().updateParents(id, xml);
+        }
+        else {
+            xml = getRestHandlerClient().updateParents(id, xml);
+        }
+        return m.unmarshalDocument(xml);
     }
 
     /**
@@ -181,19 +220,24 @@ public class OrganizationalUnitHandlerClient
      * @see de.escidoc.core.client.interfaces.OrganizationalUnitHandlerClientInterface#open(java.lang.String,
      *      de.escidoc.core.resources.common.TaskParam)
      */
+    @Override
     public Result open(final String id, final TaskParam taskParam)
         throws EscidocException, InternalClientException, TransportException {
 
-        String xml = null;
-        String taskParamString = marshalTaskParam(taskParam);
+        if (id == null)
+            throw new IllegalArgumentException("id must not be null.");
+        if (taskParam == null)
+            throw new IllegalArgumentException("taskParam must not be null.");
+
+        String xml = marshalTaskParam(taskParam);
         if (getTransport() == TransportProtocol.SOAP) {
-            xml = getSoapHandlerClient().open(id, taskParamString);
+            xml = getSoapHandlerClient().open(id, xml);
         }
         else {
-            xml = getRestHandlerClient().open(id, taskParamString);
+            xml = getRestHandlerClient().open(id, xml);
         }
         return Factory
-            .getMarshallerFactory(getTransport()).getResultMarshaller()
+            .getMarshallerFactory(getTransport()).getMarshaller(Result.class)
             .unmarshalDocument(xml);
     }
 
@@ -209,25 +253,35 @@ public class OrganizationalUnitHandlerClient
      * @see de.escidoc.core.client.interfaces.OrganizationalUnitHandlerClientInterface#close(java.lang.String,
      *      de.escidoc.core.resources.common.TaskParam)
      */
+    @Override
     public Result close(final String id, final TaskParam taskParam)
         throws EscidocException, InternalClientException, TransportException {
 
-        String xml = null;
-        String taskParamString = marshalTaskParam(taskParam);
+        if (id == null)
+            throw new IllegalArgumentException("id must not be null.");
+        if (taskParam == null)
+            throw new IllegalArgumentException("taskParam must not be null.");
+
+        String xml = marshalTaskParam(taskParam);
         if (getTransport() == TransportProtocol.SOAP) {
-            xml = getSoapHandlerClient().close(id, taskParamString);
+            xml = getSoapHandlerClient().close(id, xml);
         }
         else {
-            xml = getRestHandlerClient().close(id, taskParamString);
+            xml = getRestHandlerClient().close(id, xml);
         }
         return Factory
-            .getMarshallerFactory(getTransport()).getResultMarshaller()
+            .getMarshallerFactory(getTransport()).getMarshaller(Result.class)
             .unmarshalDocument(xml);
     }
 
-    /* (non-Javadoc)
-     * @see de.escidoc.core.client.interfaces.ResourceHandlerInterface#retrieveProperties(java.lang.String)
+    /*
+     * (non-Javadoc)
+     * 
+     * @see
+     * de.escidoc.core.client.interfaces.ResourceHandlerInterface#retrieveProperties
+     * (java.lang.String)
      */
+    @Override
     public Properties retrieveProperties(final String id)
         throws EscidocException, InternalClientException, TransportException {
 
@@ -241,8 +295,12 @@ public class OrganizationalUnitHandlerClient
      * @throws EscidocClientException
      * @see de.escidoc.core.client.interfaces.OrganizationalUnitHandlerClientInterface#delete(java.lang.String)
      */
+    @Override
     public void delete(final String id) throws EscidocException,
         InternalClientException, TransportException {
+
+        if (id == null)
+            throw new IllegalArgumentException("id must not be null.");
 
         if (getTransport() == TransportProtocol.SOAP) {
             getSoapHandlerClient().delete(id);
@@ -262,15 +320,28 @@ public class OrganizationalUnitHandlerClient
      * @throws TransportException
      * @see de.escidoc.core.client.interfaces.OrganizationalUnitHandlerClientInterface#getLastModificationDate(java.lang.String)
      */
+    @Override
     @Deprecated
     public DateTime getLastModificationDate(final String id)
         throws EscidocException, InternalClientException, TransportException {
 
-        return getSoapHandlerClient().getLastModificationDate(id);
+        if (id == null)
+            throw new IllegalArgumentException("id must not be null.");
+
+        if (getTransport() == TransportProtocol.SOAP) {
+            return getSoapHandlerClient().getLastModificationDate(id);
+        }
+        else {
+            return getRestHandlerClient().getLastModificationDate(id);
+        }
     }
-    
-    /* (non-Javadoc)
-     * @see de.escidoc.core.client.interfaces.ResourceHandlerInterface#assignObjectPid(java.lang.String, de.escidoc.core.resources.common.TaskParam)
+
+    /*
+     * (non-Javadoc)
+     * 
+     * @see
+     * de.escidoc.core.client.interfaces.ResourceHandlerInterface#assignObjectPid
+     * (java.lang.String, de.escidoc.core.resources.common.TaskParam)
      */
     @Override
     public Result assignObjectPid(final String id, final TaskParam taskParam)
@@ -287,8 +358,13 @@ public class OrganizationalUnitHandlerClient
      * @throws InternalClientException
      * @throws TransportException
      */
+    @Override
     public Parents retrieveParents(final String id) throws EscidocException,
         InternalClientException, TransportException {
+
+        if (id == null)
+            throw new IllegalArgumentException("id must not be null.");
+
         String xml = null;
         if (getTransport() == TransportProtocol.SOAP) {
             xml = getSoapHandlerClient().retrieveParents(id);
@@ -297,15 +373,25 @@ public class OrganizationalUnitHandlerClient
             xml = getRestHandlerClient().retrieveParents(id);
         }
         return Factory
-            .getMarshallerFactory(getTransport()).getParentsMarshaller()
+            .getMarshallerFactory(getTransport())
+            .getMarshaller(MarshallerFactory.CLASS_PARENTS)
             .unmarshalDocument(xml);
     }
 
-    /* (non-Javadoc)
-     * @see de.escidoc.core.client.interfaces.OrganizationalUnitHandlerClientInterface#retrieveParentObjects(java.lang.String)
+    /*
+     * (non-Javadoc)
+     * 
+     * @see
+     * de.escidoc.core.client.interfaces.OrganizationalUnitHandlerClientInterface
+     * #retrieveParentObjects(java.lang.String)
      */
+    @Override
     public OrganizationalUnitList retrieveParentObjects(final String id)
         throws EscidocException, InternalClientException, TransportException {
+
+        if (id == null)
+            throw new IllegalArgumentException("id must not be null.");
+
         String xml = null;
         if (getTransport() == TransportProtocol.SOAP) {
             xml = getSoapHandlerClient().retrieveParentObjects(id);
@@ -315,15 +401,24 @@ public class OrganizationalUnitHandlerClient
         }
         return Factory
             .getMarshallerFactory(getTransport())
-            .getOrganizationalUnitListMarshaller().unmarshalDocument(xml);
+            .getMarshaller(MarshallerFactory.CLASS_ORGANIZATIONAL_UNIT_LIST)
+            .unmarshalDocument(xml);
     }
 
-    /* (non-Javadoc)
-     * @see de.escidoc.core.client.interfaces.OrganizationalUnitHandlerClientInterface#retrieveChildObjects(java.lang.String)
+    /*
+     * (non-Javadoc)
+     * 
+     * @see
+     * de.escidoc.core.client.interfaces.OrganizationalUnitHandlerClientInterface
+     * #retrieveChildObjects(java.lang.String)
      */
     @Override
     public OrganizationalUnitList retrieveChildObjects(final String id)
         throws EscidocException, InternalClientException, TransportException {
+
+        if (id == null)
+            throw new IllegalArgumentException("id must not be null.");
+
         String xml = null;
         if (getTransport() == TransportProtocol.SOAP) {
             xml = getSoapHandlerClient().retrieveChildObjects(id);
@@ -333,7 +428,8 @@ public class OrganizationalUnitHandlerClient
         }
         return Factory
             .getMarshallerFactory(getTransport())
-            .getOrganizationalUnitListMarshaller().unmarshalDocument(xml);
+            .getMarshaller(MarshallerFactory.CLASS_ORGANIZATIONAL_UNIT_LIST)
+            .unmarshalDocument(xml);
     }
 
     /**
@@ -349,26 +445,26 @@ public class OrganizationalUnitHandlerClient
      * @throws TransportException
      *             Thrown if in case of failure on transport level.
      */
+    @Override
     @Deprecated
     public OrganizationalUnitList retrieveOrganizationalUnits(
         final TaskParam taskParam) throws EscidocException,
         InternalClientException, TransportException {
 
-        String taskParamString = marshalTaskParam(taskParam);
-        String xml = null;
+        if (taskParam == null)
+            throw new IllegalArgumentException("taskParam must not be null.");
+
+        String xml = marshalTaskParam(taskParam);
         if (getTransport() == TransportProtocol.SOAP) {
-            xml =
-                getSoapHandlerClient().retrieveOrganizationalUnits(
-                    taskParamString);
+            xml = getSoapHandlerClient().retrieveOrganizationalUnits(xml);
         }
         else {
-            xml =
-                getRestHandlerClient().retrieveOrganizationalUnits(
-                    taskParamString);
+            xml = getRestHandlerClient().retrieveOrganizationalUnits(xml);
         }
         return Factory
             .getMarshallerFactory(getTransport())
-            .getOrganizationalUnitListMarshaller().unmarshalDocument(xml);
+            .getMarshaller(MarshallerFactory.CLASS_ORGANIZATIONAL_UNIT_LIST)
+            .unmarshalDocument(xml);
     }
 
     /**
@@ -384,13 +480,14 @@ public class OrganizationalUnitHandlerClient
      * @throws TransportException
      *             Thrown if in case of failure on transport level.
      */
+    @Override
     public SearchRetrieveResponse retrieveOrganizationalUnits(
         final SearchRetrieveRequestType request) throws EscidocException,
         InternalClientException, TransportException {
 
         if (request == null)
             throw new IllegalArgumentException("request must not be null.");
-        
+
         String xml = null;
         if (getTransport() == TransportProtocol.SOAP) {
             xml = getSoapHandlerClient().retrieveOrganizationalUnits(request);
@@ -400,15 +497,22 @@ public class OrganizationalUnitHandlerClient
         }
         return Factory
             .getMarshallerFactory(getTransport())
-            .getSearchRetrieveResponseMarshaller().unmarshalDocument(xml);
+            .getMarshaller(MarshallerFactory.CLASS_SEARCH_RETRIEVE_RESPONSE)
+            .unmarshalDocument(xml);
     }
 
+    /**
+     * 
+     */
     @Override
     public Collection<OrganizationalUnit> retrieveOrganizationalUnitsAsList(
-        final SearchRetrieveRequestType filter) throws EscidocException,
+        final SearchRetrieveRequestType request) throws EscidocException,
         InternalClientException, TransportException {
 
-        SearchRetrieveResponse response = retrieveOrganizationalUnits(filter);
+        if (request == null)
+            throw new IllegalArgumentException("filter must not be null.");
+
+        SearchRetrieveResponse response = retrieveOrganizationalUnits(request);
         Collection<OrganizationalUnit> results =
             new LinkedList<OrganizationalUnit>();
 
@@ -428,7 +532,7 @@ public class OrganizationalUnitHandlerClient
     /**
      * Retrieve Organizational Units (Filter for Organizational Units).
      * 
-     * @param filter
+     * @param request
      *            Filter parameter
      * @return ExplainRecord
      * @throws EscidocException
@@ -438,20 +542,25 @@ public class OrganizationalUnitHandlerClient
      * @throws TransportException
      *             Thrown if in case of failure on transport level.
      */
+    @Override
     public ExplainResponse retrieveOrganizationalUnits(
-        final ExplainRequestType filter) throws EscidocException,
+        final ExplainRequestType request) throws EscidocException,
         InternalClientException, TransportException {
+
+        if (request == null)
+            throw new IllegalArgumentException("request must not be null.");
 
         String xml = null;
         if (getTransport() == TransportProtocol.SOAP) {
-            xml = getSoapHandlerClient().retrieveOrganizationalUnits(filter);
+            xml = getSoapHandlerClient().retrieveOrganizationalUnits(request);
         }
         else {
-            xml = getRestHandlerClient().retrieveOrganizationalUnits(filter);
+            xml = getRestHandlerClient().retrieveOrganizationalUnits(request);
         }
         return Factory
             .getMarshallerFactory(getTransport())
-            .getExplainResponseMarshaller().unmarshalDocument(xml);
+            .getMarshaller(MarshallerFactory.CLASS_EXPLAIN_RESPONSE)
+            .unmarshalDocument(xml);
     }
 
     @Override
@@ -472,4 +581,14 @@ public class OrganizationalUnitHandlerClient
         return new RestOrganizationalUnitHandlerClient(getServiceAddress());
     }
 
+    /**
+     * 
+     * @return
+     * @throws InternalClientException
+     */
+    private Marshaller<OrganizationalUnit> getMarshaller()
+        throws InternalClientException {
+        return Factory.getMarshallerFactory(getTransport()).getMarshaller(
+            MarshallerFactory.CLASS_ORGANIZATIONAL_UNIT);
+    }
 }
